Do not specify a target for  animation and skin bufferViews
The glTF spec states that for a bufferView's target property: "When this
is not provided, the bufferView contains animation or skin data." As a
result of this commit, a value of None can be used for the target of
Buffer.add_view() to create bufferViews with no target.

Thanks to lukesanantonio for bringing this to my attention and providing
some initial code.